+++ Assets/Scripts/DataTable/DataTable_NormalEnforce.cs
@@ -11,7 +11,7 @@ public class DataTable_NormalEnforce : DataTable
         public int Normal_Enforce_LV { get; set; }
         public float AttackUp { get; set; }
         public int IngredientID { get; set; }
-        public float IngredientNum { get; set; }
+        public int IngredientNum { get; set; }
         public int Gold_Cost { get; set; }
         public int Class { get; set; }
     }
